feat: 使用P环闭环完成直线_转弯_直线的基本路径运动

+++ src/utils/pid.h
@@ -25,9 +25,9 @@
  * 适用于：电机速度闭环控制（PWM 占空比增量调整）
  */
 typedef struct {
-    int16_t kp;              // 比例系数 Kp（整数表示，实际值 = kp/100）
-    int16_t ki;              // 积分系数 Ki（整数表示，实际值 = ki/100）
-    int16_t kd;              // 微分系数 Kd（整数表示，实际值 = kd/100）
+    int16_t kp;              // 比例系数 Kp（整数表示，实际值 = kp/1000）
+    int16_t ki;              // 积分系数 Ki（整数表示，实际值 = ki/1000）
+    int16_t kd;              // 微分系数 Kd（整数表示，实际值 = kd/1000）
     
     int16_t error_prev;      // 上一次误差 e(k-1)
     int16_t error_prev2;     // 上上次误差 e(k-2)
@@ -46,9 +46,9 @@ typedef struct {
  * 适用于：陀螺仪角度控制、舵机角度控制
  */
 typedef struct {
-    int16_t kp;              // 比例系数 Kp（整数表示，实际值 = kp/100）
-    int16_t ki;              // 积分系数 Ki（整数表示，实际值 = ki/100）
-    int16_t kd;              // 微分系数 Kd（整数表示，实际值 = kd/100）
+    int16_t kp;              // 比例系数 Kp（整数表示，实际值 = kp/1000）
+    int16_t ki;              // 积分系数 Ki（整数表示，实际值 = ki/1000）
+    int16_t kd;              // 微分系数 Kd（整数表示，实际值 = kd/1000）
     
     int16_t error_prev;      // 上一次误差 e(k-1)
     int32_t integral_sum;    // 积分累加值 ∑e(i)
@@ -72,7 +72,9 @@ typedef struct {
  * @param output_min: 输出限幅下限
  */
 void PID_Increment_Init(PID_Increment_t *pid, int16_t kp, int16_t ki, int16_t kd,
-                        int16_t output_max, int16_t output_min);
+                         int16_t output_max, int16_t output_min);
+
+/** 注：增量式 PID 系数仍使用 /100 缩放 */
 
 /**
  * @brief 计算增量式 PID 输出
@@ -94,9 +96,9 @@ void PID_Increment_Reset(PID_Increment_t *pid);
 /**
  * @brief 初始化位置式 PID 控制器
  * @param pid: PID 控制器指针
- * @param kp: 比例系数（整数表示，实际值 = kp/100）
- * @param ki: 积分系数（整数表示，实际值 = ki/100）
- * @param kd: 微分系数（整数表示，实际值 = kd/100）
+ * @param kp: 比例系数（整数表示，实际值 = kp/1000）
+ * @param ki: 积分系数（整数表示，实际值 = ki/1000）
+ * @param kd: 微分系数（整数表示，实际值 = kd/1000）
  * @param integral_max: 积分限幅上限
  * @param integral_min: 积分限幅下限
  * @param output_max: 输出限幅上限
@@ -113,7 +115,7 @@ void PID_Position_Init(PID_Position_t *pid, int16_t kp, int16_t ki, int16_t kd,
  * @param feedback: 反馈值（实际值）
  * @return 绝对输出值 u(k)
  */
-int16_t PID_Position_Calculate(PID_Position_t *pid, int16_t setpoint, int16_t feedback);
+int32_t PID_Position_Calculate(PID_Position_t *pid, int32_t setpoint, int32_t feedback);
 
 /**
  * @brief 重置位置式 PID 控制器
